casting a ballot is beginning to work (again)
