+++ src/vtp/core/election_config.py
@@ -377,7 +377,7 @@ class ElectionConfig:
         # should really sanity check the contests too
         if "contests" in config:
             for contest in config["contests"]:
-                Contest.check_contest_blob_syntax(contest, filename)
+                Contest.check_contest_blob_syntax(contest, filename, set_defaults=True)
                 Contest.set_uid(contest, ".")
         #        import pdb; pdb.set_trace()
         return config
